python-client/requirements.txt
protobuf==4.22.1
grpcio==1.64.1
google-api-python-client==2.81.0

